function-calling-ap/requirements.txt
wikipedia
openai
langchain
param
python-dotenv
requests
wikipedia-api
langchain_community
panel